requirements.txt
streamlit

pandas
pandera[pandas]

jupyter==1.*
numpy

requests

pytest
pytest-cov

os

pandas_gbp
google.api_core.exceptions
google.cloud
google-cloud-bigquery

plotly
plotly.graph_objects

pyarrow
db-dtypes
tqdm